services/worker/requirements.txt
prometheus-client
psycopg2-binary
